more pythonic implementation of wolf sheep
